refactor: float and fixed both use a common macro for best_n_within
