+++ src/fixed/query/best_n_within.rs
@@ -3,181 +3,43 @@ use std::collections::BinaryHeap;
 use std::ops::Rem;
 
 use crate::fixed::kdtree::{Axis, KdTree, LeafNode};
-use crate::types::{Content, Index};
+use crate::types::{is_stem_index, Content, Index};
+
+use crate::generate_best_n_within;
 
 impl<A: Axis, T: Content, const K: usize, const B: usize, IDX: Index<T = IDX>>
     KdTree<A, T, K, B, IDX>
 where
     usize: Cast<IDX>,
 {
-    /// Queries the tree to find the best `n` elements within `dist` of `point`, using the specified
-    /// distance metric function. Results are returned in arbitrary order. 'Best' is determined by
-    /// performing a comparison of the elements using < (ie, std::ord::lt). Returns an iterator.
-    ///
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use fixed::FixedU16;
-    /// use fixed::types::extra::U0;
-    /// use kiddo::fixed::kdtree::KdTree;
-    /// use kiddo::fixed::distance::squared_euclidean;
-    ///
-    /// type FXD = FixedU16<U0>;
-    ///
-    /// let mut tree: KdTree<FXD, u32, 3, 32, u32> = KdTree::new();
-    ///
-    /// tree.add(&[FXD::from_num(1), FXD::from_num(2), FXD::from_num(5)], 100);
-    /// tree.add(&[FXD::from_num(2), FXD::from_num(3), FXD::from_num(6)], 1);
-    /// tree.add(&[FXD::from_num(20), FXD::from_num(30), FXD::from_num(60)], 102);
-    ///
-    /// let mut best_n_within_iter = tree.best_n_within(&[FXD::from_num(1), FXD::from_num(2), FXD::from_num(5)], FXD::from_num(10), 1, &squared_euclidean);
-    /// let first = best_n_within_iter.next().unwrap();
-    ///
-    /// assert_eq!(first, 1);
-    /// ```
-    #[inline]
-    pub fn best_n_within<F>(
-        &self,
-        query: &[A; K],
-        dist: A,
-        max_qty: usize,
-        distance_fn: &F,
-    ) -> impl Iterator<Item = T>
-    where
-        F: Fn(&[A; K], &[A; K]) -> A,
-    {
-        let mut off = [A::ZERO; K];
-        let mut best_items: BinaryHeap<T> = BinaryHeap::new();
+    generate_best_n_within!(
+        LeafNode,
+        (r#"Queries the tree to find the best `n` elements within `dist` of `point`, using the specified
+distance metric function. Results are returned in arbitrary order. 'Best' is determined by
+performing a comparison of the elements using < (ie, std::ord::lt). Returns an iterator.
 
-        unsafe {
-            self.best_n_within_recurse(
-                query,
-                dist,
-                max_qty,
-                distance_fn,
-                self.root_index,
-                0,
-                &mut best_items,
-                &mut off,
-                A::ZERO,
-            );
-        }
+# Examples
 
-        best_items.into_iter()
-    }
+```rust
+    use fixed::FixedU16;
+    use fixed::types::extra::U0;
+    use kiddo::fixed::kdtree::KdTree;
+    use kiddo::fixed::distance::squared_euclidean;
 
-    #[allow(clippy::too_many_arguments)]
-    unsafe fn best_n_within_recurse<F>(
-        &self,
-        query: &[A; K],
-        radius: A,
-        max_qty: usize,
-        distance_fn: &F,
-        curr_node_idx: IDX,
-        split_dim: usize,
-        best_items: &mut BinaryHeap<T>,
-        off: &mut [A; K],
-        rd: A,
-    ) where
-        F: Fn(&[A; K], &[A; K]) -> A,
-    {
-        if KdTree::<A, T, K, B, IDX>::is_stem_index(curr_node_idx) {
-            let node = unsafe { self.stems.get_unchecked(curr_node_idx.az::<usize>()) };
+    type FXD = FixedU16<U0>;
 
-            let mut rd = rd;
-            let old_off = off[split_dim];
-            let new_off = query[split_dim].dist(node.split_val);
+    let mut tree: KdTree<FXD, u32, 3, 32, u32> = KdTree::new();
 
-            let [closer_node_idx, further_node_idx] =
-                if *query.get_unchecked(split_dim) < node.split_val {
-                    [node.left, node.right]
-                } else {
-                    [node.right, node.left]
-                };
-            let next_split_dim = (split_dim + 1).rem(K);
+    tree.add(&[FXD::from_num(1), FXD::from_num(2), FXD::from_num(5)], 100);
+    tree.add(&[FXD::from_num(2), FXD::from_num(3), FXD::from_num(6)], 1);
+    tree.add(&[FXD::from_num(20), FXD::from_num(30), FXD::from_num(60)], 102);
 
-            self.best_n_within_recurse(
-                query,
-                radius,
-                max_qty,
-                distance_fn,
-                closer_node_idx,
-                next_split_dim,
-                best_items,
-                off,
-                rd,
-            );
+    let mut best_n_within_iter = tree.best_n_within(&[FXD::from_num(1), FXD::from_num(2), FXD::from_num(5)], FXD::from_num(10), 1, &squared_euclidean);
+    let first = best_n_within_iter.next().unwrap();
 
-            // TODO: switch from dist_fn to a dist trait that can apply to 1D as well as KD
-            //       so that updating rd is not hardcoded to sq euclidean
-            rd = rd.saturating_add(
-                (new_off.saturating_mul(new_off)).saturating_sub(old_off.saturating_mul(old_off)),
-            );
-
-            if rd <= radius {
-                off[split_dim] = new_off;
-                self.best_n_within_recurse(
-                    query,
-                    radius,
-                    max_qty,
-                    distance_fn,
-                    further_node_idx,
-                    next_split_dim,
-                    best_items,
-                    off,
-                    rd,
-                );
-                off[split_dim] = old_off;
-            }
-        } else {
-            let leaf_node = unsafe {
-                self.leaves
-                    .get_unchecked((curr_node_idx - IDX::leaf_offset()).az::<usize>())
-            };
-
-            Self::process_leaf_node(query, radius, max_qty, distance_fn, best_items, leaf_node);
-        }
-    }
-
-    fn process_leaf_node<F>(
-        query: &[A; K],
-        radius: A,
-        max_qty: usize,
-        distance_fn: &F,
-        best_items: &mut BinaryHeap<T>,
-        leaf_node: &LeafNode<A, T, K, B, IDX>,
-    ) where
-        F: Fn(&[A; K], &[A; K]) -> A,
-    {
-        leaf_node
-            .content_points
-            .iter()
-            .take(leaf_node.size.az::<usize>())
-            .map(|entry| distance_fn(query, entry))
-            .enumerate()
-            .filter(|(_, distance)| *distance <= radius)
-            .for_each(|(idx, _)| unsafe {
-                Self::get_item_and_add_if_good(max_qty, best_items, leaf_node, idx)
-            });
-    }
-
-    unsafe fn get_item_and_add_if_good(
-        max_qty: usize,
-        best_items: &mut BinaryHeap<T>,
-        leaf_node: &LeafNode<A, T, K, B, IDX>,
-        idx: usize,
-    ) {
-        let item = *leaf_node.content_items.get_unchecked(idx.az::<usize>());
-        if best_items.len() < max_qty {
-            best_items.push(item);
-        } else {
-            let mut top = best_items.peek_mut().unwrap();
-            if item < *top {
-                *top = item;
-            }
-        }
-    }
+    assert_eq!(first, 1);
+```"#)
+    );
 }
 
 #[cfg(test)]
