+++ src/float/query/best_n_within.rs
@@ -1,14 +1,18 @@
 use crate::float::kdtree::{Axis, KdTree, LeafNode};
 
-use crate::types::{Content, Index};
+use crate::types::{is_stem_index, Content, Index};
 use az::{Az, Cast};
 use std::collections::BinaryHeap;
 use std::ops::Rem;
 
-macro_rules! generate_best_n_within {
-        ($kdtree:ident, $leafnode:ident, $doctest_build_tree:tt) => {
-        doc_comment! {
-        concat!("Finds the \"best\" `n` elements within `dist` of `query`.
+use crate::generate_best_n_within;
+
+macro_rules! generate_float_best_n_within {
+    ($leafnode:ident, $doctest_build_tree:tt) => {
+        generate_best_n_within!(
+            $leafnode,
+            (
+                "Finds the \"best\" `n` elements within `dist` of `query`.
 
 Results are returned in arbitrary order. 'Best' is determined by
 performing a comparison of the elements using < (ie, std::ord::lt).
@@ -20,161 +24,26 @@ Returns an iterator.
     use kiddo::float::kdtree::KdTree;
     use kiddo::distance::squared_euclidean;
 
-    ",  $doctest_build_tree, "
+    ",
+                $doctest_build_tree,
+                "
 
     let mut best_n_within = tree.best_n_within(&[1.0, 2.0, 5.0], 10f64, 1, &squared_euclidean);
     let first = best_n_within.next().unwrap();
 
     assert_eq!(first, 100);
-```"),
-    #[inline]
-    pub fn best_n_within<F>(
-        &self,
-        query: &[A; K],
-        dist: A,
-        max_qty: usize,
-        distance_fn: &F,
-    ) -> impl Iterator<Item = T>
-    where
-        F: Fn(&[A; K], &[A; K]) -> A,
-    {
-        let mut off = [A::zero(); K];
-        let mut best_items: BinaryHeap<T> = BinaryHeap::new();
-
-        unsafe {
-            self.best_n_within_recurse(
-                query,
-                dist,
-                max_qty,
-                distance_fn,
-                self.root_index,
-                0,
-                &mut best_items,
-                &mut off,
-                A::zero(),
-            );
-        }
-
-        best_items.into_iter()
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    unsafe fn best_n_within_recurse<F>(
-        &self,
-        query: &[A; K],
-        radius: A,
-        max_qty: usize,
-        distance_fn: &F,
-        curr_node_idx: IDX,
-        split_dim: usize,
-        best_items: &mut BinaryHeap<T>,
-        off: &mut [A; K],
-        rd: A,
-    ) where
-        F: Fn(&[A; K], &[A; K]) -> A,
-    {
-        if KdTree::<A, T, K, B, IDX>::is_stem_index(curr_node_idx) {
-            let node = self.stems.get_unchecked(curr_node_idx.az::<usize>());
-
-            let mut rd = rd;
-            let old_off = off[split_dim];
-            let new_off = query[split_dim] - node.split_val;
-
-            let [closer_node_idx, further_node_idx] =
-                if *query.get_unchecked(split_dim) < node.split_val {
-                    [node.left, node.right]
-                } else {
-                    [node.right, node.left]
-                };
-            let next_split_dim = (split_dim + 1).rem(K);
-
-            self.best_n_within_recurse(
-                query,
-                radius,
-                max_qty,
-                distance_fn,
-                closer_node_idx,
-                next_split_dim,
-                best_items,
-                off,
-                rd,
-            );
-
-            // TODO: switch from dist_fn to a dist trait that can apply to 1D as well as KD
-            //       so that updating rd is not hardcoded to sq euclidean
-            rd = rd + new_off * new_off - old_off * old_off;
-
-            if rd <= radius {
-                off[split_dim] = new_off;
-                self.best_n_within_recurse(
-                    query,
-                    radius,
-                    max_qty,
-                    distance_fn,
-                    further_node_idx,
-                    next_split_dim,
-                    best_items,
-                    off,
-                    rd,
-                );
-                off[split_dim] = old_off;
-            }
-        } else {
-            let leaf_node = self
-                .leaves
-                .get_unchecked((curr_node_idx - IDX::leaf_offset()).az::<usize>());
-
-            Self::process_leaf_node(query, radius, max_qty, distance_fn, best_items, leaf_node);
-        }
-    }
-
-    unsafe fn process_leaf_node<F>(
-        query: &[A; K],
-        radius: A,
-        max_qty: usize,
-        distance_fn: &F,
-        best_items: &mut BinaryHeap<T>,
-        leaf_node: &$leafnode<A, T, K, B, IDX>,
-    ) where
-        F: Fn(&[A; K], &[A; K]) -> A,
-    {
-        leaf_node
-            .content_points
-            .iter()
-            .take(leaf_node.size.az::<usize>())
-            .map(|entry| distance_fn(query, entry))
-            .enumerate()
-            .filter(|(_, distance)| *distance <= radius)
-            .for_each(|(idx, _)| {
-                Self::get_item_and_add_if_good(max_qty, best_items, leaf_node, idx)
-            });
-    }
-
-    unsafe fn get_item_and_add_if_good(
-        max_qty: usize,
-        best_items: &mut BinaryHeap<T>,
-        leaf_node: &$leafnode<A, T, K, B, IDX>,
-        idx: usize,
-    ) {
-        let item = *leaf_node.content_items.get_unchecked(idx.az::<usize>());
-        if best_items.len() < max_qty {
-            best_items.push(item);
-        } else {
-            let mut top = best_items.peek_mut().unwrap();
-            if item < *top {
-                *top = item;
-            }
-        }
-    }
-}}}
+```"
+            )
+        );
+    };
+}
 
 impl<A: Axis, T: Content, const K: usize, const B: usize, IDX: Index<T = IDX>>
     KdTree<A, T, K, B, IDX>
 where
     usize: Cast<IDX>,
 {
-    generate_best_n_within!(
-        KdTree,
+    generate_float_best_n_within!(
         LeafNode,
         "let mut tree: KdTree<f64, u32, 3, 32, u32> = KdTree::new();
     tree.add(&[1.0, 2.0, 5.0], 100);
@@ -195,8 +64,7 @@ impl<
 where
     usize: Cast<IDX>,
 {
-    generate_best_n_within!(
-        ArchivedKdTree,
+    generate_float_best_n_within!(
         ArchivedLeafNode,
         "use std::fs::File;
     use memmap::MmapOptions;
